sql-agent/requirements.txt
mysql-connector-python==9.0.0
openai
python-dotenv
pydo
